Added keyboard translator and fixed MakeFile

+++ kernel/src/Math.h
@@ -4,7 +4,7 @@
 
 struct Point{
 
-	uint32_t X;
-	uint32_t Y;
+	long X;
+	long Y;
 
 };
